feat: finishing GetParticipantReferenceData function
