+++ application/CohortManager/src/Functions/Functions.sln
@@ -167,7 +167,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CheckParticipantExistsTests", "..\..\..\..\tests\UnitTests\ParticipantManagementServicesTests\CheckParticipantExistsTests\CheckParticipantExistsTests.csproj", "{55EA363F-FCE2-4EA2-8CEF-02996B45C2B6}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CheckParticipantExists", "ParticipantManagementServices\CheckParticipantExists\CheckParticipantExists.csproj", "{033D3F4B-81E6-45DF-B457-F8A7C9BDFA23}"
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TransformReasonForRemovalTests", "..\..\..\..\tests\UnitTests\TransformDataServiceTests\TransformReasonForRemovalTests\TransformReasonForRemovalTests.csproj", "{C31490BE-8D6F-4F0B-ABC7-9D3B97FE20C4}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TransformDataServiceTests", "..\..\..\..\tests\UnitTests\TransformDataServiceTests\TransformDataServiceTests\TransformDataServiceTests.csproj", "{DCBF3583-DC87-46BE-93DC-C345E779D95C}"
 EndProject
@@ -180,6 +179,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ValidateDatesTests", "..\..\..\..\tests\UnitTests\CaasIntegrationTests\ValidateDates\ValidateDatesTests.csproj", "{9005A919-F64D-4723-87A7-595BB4E64563}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataServiceTests", "..\..\..\..\tests\UnitTests\DataServiceTests\DataServiceTests.csproj", "{E26B2D73-BD60-4AD8-B403-939ACEDE9429}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GetParticipantReferenceData", "ParticipantManagementServices\GetParticipantReferenceData\GetParticipantReferenceData.csproj", "{DCBEE667-B229-462D-A245-FC9EE4A58A1F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -477,6 +478,18 @@ Global
 		{E26B2D73-BD60-4AD8-B403-939ACEDE9429}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E26B2D73-BD60-4AD8-B403-939ACEDE9429}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E26B2D73-BD60-4AD8-B403-939ACEDE9429}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DCBEE667-B229-462D-A245-FC9EE4A58A1F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DCBEE667-B229-462D-A245-FC9EE4A58A1F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DCBEE667-B229-462D-A245-FC9EE4A58A1F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{DCBEE667-B229-462D-A245-FC9EE4A58A1F}.Debug|x64.Build.0 = Debug|Any CPU
+		{DCBEE667-B229-462D-A245-FC9EE4A58A1F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{DCBEE667-B229-462D-A245-FC9EE4A58A1F}.Debug|x86.Build.0 = Debug|Any CPU
+		{DCBEE667-B229-462D-A245-FC9EE4A58A1F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DCBEE667-B229-462D-A245-FC9EE4A58A1F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DCBEE667-B229-462D-A245-FC9EE4A58A1F}.Release|x64.ActiveCfg = Release|Any CPU
+		{DCBEE667-B229-462D-A245-FC9EE4A58A1F}.Release|x64.Build.0 = Release|Any CPU
+		{DCBEE667-B229-462D-A245-FC9EE4A58A1F}.Release|x86.ActiveCfg = Release|Any CPU
+		{DCBEE667-B229-462D-A245-FC9EE4A58A1F}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -522,6 +535,7 @@ Global
 		{9B9E8E86-601F-4C10-B6EF-348AF9A168AA} = {3A61DBD3-B2D4-45D7-9CB6-9BC4C9586467}
 		{D8990183-52BA-4A1F-B333-5B32AAF14348} = {3A61DBD3-B2D4-45D7-9CB6-9BC4C9586467}
 		{28F22CA5-7CE5-4D2E-8E0E-AD1D9A11C0C8} = {3A61DBD3-B2D4-45D7-9CB6-9BC4C9586467}
+		{DCBEE667-B229-462D-A245-FC9EE4A58A1F} = {E1ADFCCC-1FF7-40E6-8D82-301F3FA909A7}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6EACF2AC-4BC4-46CE-B6EB-1556257C4C21}
