+++ application/CohortManager/src/Functions/ParticipantManagementServices/GetParticipantReferenceData/Dockerfile
@@ -1,6 +1,6 @@
 FROM mcr.microsoft.com/dotnet/sdk:8.0 AS installer-env
 
-COPY ./screeningDataServices/updateParticipantDetails /src/dotnet-function-app
+COPY ./ParticipantManagementServices/GetParticipantReferenceData /src/dotnet-function-app
 COPY ./Shared /Shared
 RUN cd /src/dotnet-function-app && \
 mkdir -p /home/site/wwwroot && \
