+++ application/CohortManager/src/Functions/ParticipantManagementServices/GetParticipantReferenceData/Program.cs
@@ -1,24 +1,23 @@
 using Common;
-using Common.Interfaces;
 using Data.Database;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using Model;
+using DataServices.Client;
+
 
 var host = new HostBuilder()
     .ConfigureFunctionsWorkerDefaults()
     .AddDataServicesHandler()
-        .AddDataService<ParticipantManagement>(Environment.GetEnvironmentVariable("ParticipantManagementUrl"))
-        .AddDataService<ParticipantManagement>(Environment.GetEnvironmentVariable("ParticipantManagementUrl"))
-        .AddDataService<ParticipantManagement>(Environment.GetEnvironmentVariable("ParticipantManagementUrl"))
+        .AddDataService<HigherRiskReferralReasonLkp>(Environment.GetEnvironmentVariable("HigherRiskReferralReasonLkpDataServiceUrl"))
+        .AddDataService<GeneCodeLkp>(Environment.GetEnvironmentVariable("GeneCodeLkpDataServiceUrl"))
         .Build()
     .ConfigureServices(services =>
     {
         services.AddSingleton<ICreateResponse, CreateResponse>();
         services.AddSingleton<IDatabaseHelper, DatabaseHelper>();
-        services.AddSingleton<ICallFunction, CallFunction>();
     })
     .AddDatabaseConnection()
-    .AddExceptionHandler()
     .Build();
 
 host.Run();
